pyproject.toml
[project]
name = "incident-management"
version = "0.1.0"
description = "Incident management with ticket auto classify"
authors = [
    {name = "Suender Oliveira"}
]
license = {text = "APACHE 2.0"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
